train function of sentEmbedding is implemented
